[fpu_irq] fix stack alignment in excp_early.S
We were encountering a problem on baremetal where some XMM
related stack operations would cause a GPF. SSE instructions
must be 16 byte aligned, and we were pushing an 8 byte value
onto the stack before entering the irq handler. This is against
the amd64 ABI, so we feel bad :^\.

This doesn't show itself on QEMU's TCG because it's too exensive
to check stack alignment in software on every memory op.

Hopefull this fixes the other problem